db_simulation/requirements.txt
pyrebase4==4.5.0
